+++ example/lib/main.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_stripe/flutter_stripe.dart';
 import 'package:stripe_example/.env.dart';
-import 'screens/checkout/checkout_screen.dart';
 import 'screens/screens.dart';
 import 'widgets/dismiss_focus_overlay.dart';
 
@@ -22,15 +21,7 @@ class App extends StatelessWidget {
     return DismissFocusOverlay(
       child: MaterialApp(
         theme: exampleAppTheme,
-        routes: {...CheckoutScreenExample.routes},
-        onGenerateInitialRoutes: (settings) {
-          print(settings);
-          return [
-            MaterialPageRoute(builder: (context) {
-              return HomePage();
-            })
-          ];
-        },
+        home: HomePage(),
         navigatorObservers: [],
       ),
     );