+++ packages/stripe_checkout/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.2
+
+- `successUrl` and `canceledUrl` are now required when using checkout with a webview and they should use https
+- Fix [#576]: `Live Stripe.js integrations must use HTTPS error`
+
 ## 0.1.1
 
 - Fix [#494] Make `ShippingDetails` parameters nullable